Start porting to intrinsics.rs
